requirements.txt
pandas
mysql-connector-python
flask
langchain_openai
langchain
python-dotenv